Build a Rake task to run tests within a container
This commit refactors the structure of the docker related rake tasks to
make sure they always use the same running instance of the container.

It provides a rake task `rake docker:tests` that runs the tests within
the most recently created and running container.

It also provides some convenience methods to kill all the containers